renaming and re-arranging
