+++ pkg/yamlParser/reader.go
@@ -1,4 +1,4 @@
-package hulak_yaml_reader
+package yamlParser
 
 import (
 	"encoding/json"
@@ -7,31 +7,7 @@ import (
 	"os"
 
 	yaml "github.com/goccy/go-yaml"
-	utils "github.com/xaaha/hulak/pkg/utils"
 )
-
-type User struct {
-	Name  string `yaml:"name"`
-	Age   string `yaml:"age"`
-	Email string `yaml:"email"`
-}
-
-type GraphQl struct {
-	Variable map[string]interface{}
-	Query    string
-}
-
-// type of Body in a yaml file
-// binary type is not yet configured
-// only one is possible that could be passed
-type Body struct {
-	Graphql            GraphQl
-	RawString          string
-	FormData           []utils.KeyValuePair
-	UrlEncodedFormData []utils.KeyValuePair
-}
-
-// type Url struct{}
 
 func ReadingYamlWithStruct() {
 	file, err := os.Open("test_collection/user.yaml")
